RFC: Remove to helper.py
