+++ skorch/tests/test_utils.py
@@ -417,91 +417,3 @@ class TestIsSkorchDataset:
         type_truth_table())
     def test_data_types(self, is_skorch_dataset, input_data, expected):
         assert is_skorch_dataset(input_data) == expected
-
-
-class TestFilterRequiresGrad():
-
-    @pytest.fixture
-    def filter_requires_grad(self):
-        from skorch.utils import filter_requires_grad
-        return filter_requires_grad
-
-    def test_all_parameters_requires_gradient(
-            self, filter_requires_grad):
-        pgroups = [{
-            'params': [torch.zeros(1, requires_grad=True),
-                       torch.zeros(1, requires_grad=True)],
-            'lr': 0.1
-        }, {
-            'params': [torch.zeros(1, requires_grad=True)]
-        }]
-
-        filter_pgroups = list(filter_requires_grad(pgroups))
-        assert len(filter_pgroups) == 2
-        assert len(list(filter_pgroups[0]['params'])) == 2
-        assert len(list((filter_pgroups[1]['params']))) == 1
-
-        assert filter_pgroups[0]['lr'] == 0.1
-
-    def test_some_params_requires_gradient(
-            self, filter_requires_grad):
-        pgroups = [{
-            'params': [
-                torch.zeros(1, requires_grad=True),
-                torch.zeros(1, requires_grad=False)
-            ], 'lr':0.1
-        }, {
-            'params': [torch.zeros(1, requires_grad=False)]
-        }]
-
-        filter_pgroups = list(filter_requires_grad(pgroups))
-        assert len(filter_pgroups) == 2
-        assert len(list(filter_pgroups[0]['params'])) == 1
-        assert len(list(filter_pgroups[1]['params'])) == 0
-
-        assert filter_pgroups[0]['lr'] == 0.1
-
-    def test_does_not_drop_group_when_requires_grad_is_false(
-            self, filter_requires_grad):
-        pgroups = [{
-            'params': [
-                torch.zeros(1, requires_grad=False),
-                torch.zeros(1, requires_grad=False)
-            ], 'lr':0.1
-        }, {
-            'params': [torch.zeros(1, requires_grad=False)]
-        }]
-
-        filter_pgroups = list(filter_requires_grad(pgroups))
-        assert len(filter_pgroups) == 2
-        assert len(list(filter_pgroups[0]['params'])) == 0
-        assert len(list(filter_pgroups[1]['params'])) == 0
-
-        assert filter_pgroups[0]['lr'] == 0.1
-
-
-class TestFilteredOptimizer:
-
-    @pytest.fixture
-    def filtered_optimizer(self):
-        from skorch.utils import filtered_optimizer
-        return filtered_optimizer
-
-    def test_passes_filtered_cgroups(self, filtered_optimizer):
-        pgroups = [{
-            'params': [torch.zeros(1, requires_grad=True),
-                       torch.zeros(1, requires_grad=False)],
-            'lr': 0.1
-        }, {
-            'params': [torch.zeros(1, requires_grad=True)]
-        }]
-
-        filtered_func = filtered_optimizer(torch.optim.SGD)
-        filtered_opt = filtered_func(pgroups, lr=0.2)
-
-        assert isinstance(filtered_opt, torch.optim.SGD)
-        assert len(list(filtered_opt.param_groups[0]['params'])) == 1
-        assert len(list(filtered_opt.param_groups[1]['params'])) == 1
-
-        assert filtered_opt.param_groups[0]['lr'] == 0.1
-        assert filtered_opt.param_groups[1]['lr'] == 0.2
